Create ManagerMixin
