+++ project/common/models.py
@@ -3,6 +3,7 @@ from django.db import models
 from django.utils.translation import gettext_lazy as _
 from django_countries import fields
 from django_enum import TextChoices
+from user.models import User
 
 
 class MainInformationMixin(models.Model):
@@ -45,3 +46,34 @@ class CarInformationMixin(models.Model):
 
     class Meta:
         abstract = True
+
+
+class ModelManagerMixin(models.Manager):
+    def create_instance(self, username, email, password, **kwargs):
+        user = User.objects.create_user(
+            username=username, email=email, password=password
+        )
+        kwargs["user"] = user
+        return self.create(**kwargs)
+
+    def update_instance(self, id, **kwargs):
+        self.filter(id=id).update(**kwargs)
+        return self.get(id=id)
+
+
+class ModelCarManagerMixin(models.Manager):
+    def create_instance(self, **kwargs):
+        return self.create(**kwargs)
+
+    def update_instance(self, id, **kwargs):
+        self.filter(id=id).update(**kwargs)
+        return self.get(id=id)
+
+
+class ModelOfferManagerMixin(models.Manager):
+    def create_instance(self, **kwargs):
+        return self.create(**kwargs)
+
+    def update_instance(self, id, **kwargs):
+        self.filter(id=id).update(**kwargs)
+        return self.get(id=id)
